test(api): implement customizable api tests
