+++ packages/api/src/index.js
@@ -2,6 +2,7 @@ import express, { Router } from 'express'
 import 'express-async-errors' // it monkeypatches, i guess
 import morgan from 'morgan'
 import { json as jsonParser } from 'body-parser'
+import * as bodyParser from 'body-parser'
 import { LevelStore, PostgresStore } from './store'
 import { healthCheck } from './middleware'
 import logger from './logger'
@@ -9,6 +10,8 @@ import schema from './schema'
 import * as controllers from './controllers'
 import streamProxy from './controllers/stream-proxy'
 import * as k8s from '@kubernetes/client-node'
+
+// console.log(bodyParser)
 
 export default async function makeApp({
   storage,
@@ -42,6 +45,7 @@ export default async function makeApp({
     next()
   })
 
+  // Populate Kubernetes stuff if present
   if (kubeNamespace && (kubeBroadcasterService || kubeOrchestratorService)) {
     const kc = new k8s.KubeConfig()
     kc.loadFromDefault()
